+++ test/functs/test-func-cb-empty.js
@@ -1,0 +1,6 @@
+/*
+ * Lambda function callback equal empty string.
+ */
+exports.handler = function(event, context, callback) {
+    callback('');
+};