+++ test/functs/test-func-cb-fail.js
@@ -1,0 +1,6 @@
+/*
+ * Lambda function callback failed.
+ */
+exports.handler = function(event, context, callback) {
+    callback('Failed !');
+};