+++ test/test.js
@@ -100,7 +100,7 @@ describe("- Testing lambdalocal.js", function () {
                 environment: {
                     "envkey1": "Environment",
                     "envkey2": {"k":"v"},
-                    "envkey3": 123
+                    "envkey3": 123,
                 },
                 envfile: path.join(__dirname, "./other/env"),
                 verboseLevel: 1
@@ -117,11 +117,40 @@ describe("- Testing lambdalocal.js", function () {
                 assert.equal(process.env.envkey5, 'bar');
             });
         });
+        describe("# Environment Variables (destroy)", function () {
+            var done, err;
+            before(function (cb) {
+                var lambdalocal = require("../lib/lambdalocal.js");
+                lambdalocal.setLogger(winston);
+                lambdalocal.execute({
+                    event: require(path.join(__dirname, "./events/test-event.js")),
+                    lambdaPath: path.join(__dirname, "./functs/test-func.js"),
+                    lambdaHandler: functionName,
+                    profilePath: path.join(__dirname, "./other/debug.aws"),
+                    callbackWaitsForEmptyEventLoop: false,
+                    timeoutMs: timeoutMs,
+                    callback: function (_err, _done) {
+                        err = _err;
+                        done = _done;
+                        cb();
+                    },
+                    environment: {
+                        "isnetestlambda": "I should not exist",
+                    },
+                    envdestroy: true,
+                    envfile: path.join(__dirname, "./other/env"),
+                    verboseLevel: 1
+                });
+            });
+            it("environment should have been deleted", function () {
+                assert.equal(!("isnetestlambda" in process.env), true);
+            });
+        });
 
 
         describe("# AWS credentials", function () {
             it("should return correct credentials", function () {
-                assert.equal(process.env.AWS_DEFAULT_REGION, "not-us-east");
+                assert.equal(process.env.AWS_REGION, "not-us-east");
                 assert.equal(process.env.AWS_ACCESS_KEY_ID, "AKIAIOSFODNN7EXAMPLE");
                 assert.equal(process.env.AWS_SECRET_ACCESS_KEY, "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY");
                 assert.equal(process.env.AWS_SESSION_TOKEN, "TOKEN44545");
@@ -198,8 +227,28 @@ describe("- Testing lambdalocal.js", function () {
                         },
                         verboseLevel: 1
                     }, utils.TimeoutError);
-                })
+                });
                 cb();
+            });
+        });
+        describe("* Return Error object", function () {
+            it("should convert it to correct JSON format", function (cb) {
+                var lambdalocal = require("../lib/lambdalocal.js");
+                lambdalocal.setLogger(winston);
+                var lambdaFunc = require("./functs/test-func-cb-error.js");
+                lambdalocal.execute({
+                    event: require(path.join(__dirname, "./events/test-event.js")),
+                    lambdaFunc: lambdaFunc,
+                    lambdaHandler: functionName,
+                    callbackWaitsForEmptyEventLoop: false,
+                    timeoutMs: 1000,
+                    callback: function (err, _done) {
+                        assert.equal(err.errorType, "Error");
+                        assert.equal(err.errorMessage, "Failed for an unknown reason !");
+                        cb();
+                    },
+                    verboseLevel: 1
+                });
             });
         });
     });
@@ -241,15 +290,72 @@ describe("- Testing bin/lambda-local", function () {
             console.log(r.output.toString('utf8'));
             console.log(r.stderr.toString('utf8'));
         });
+
+        it("should end normally: callback null", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-cb-null.js -e ./events/test-event.js");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 0);
+            console.log(r.output.toString('utf8'));
+            console.log(r.stderr.toString('utf8'));
+        });
+
+        it("should end normally: callback undefined", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-cb-undefined.js -e ./events/test-event.js");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 0);
+            console.log(r.output.toString('utf8'));
+            console.log(r.stderr.toString('utf8'));
+        });
     });
+
     describe("* Failing Run", function () {
-        it("should fail", function () {
+        it("should fail: context", function () {
             var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-fail.js -e ./events/test-event.js");
             var r = spawnSync(command[0], command[1]);
             assert.equal(r.status, 1);
             console.log(r.output.toString('utf8'));
         });
+
+        it("should fail: callback string", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-cb-fail.js -e ./events/test-event.js");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 1);
+            console.log(r.output.toString('utf8'));
+        });
+
+        it("should fail: callback empty string", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-cb-empty.js -e ./events/test-event.js");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 1);
+            console.log(r.output.toString('utf8'));
+        });
+
+        it("should fail: callback false", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-cb-false.js -e ./events/test-event.js");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 1);
+            console.log(r.output.toString('utf8'));
+        });
+
+        it("should fail: callback 0", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-cb-0.js -e ./events/test-event.js");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 1);
+            console.log(r.output.toString('utf8'));
+        });
     });
+
+    describe("* Environment test run", function () {
+        it("event should have used ENV while building", function () {
+            var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-env.js -e ./events/test-event-env.js -v 1 --envdestroy -E {\"TEST_HUBID\":\"potato\"}");
+            var r = spawnSync(command[0], command[1]);
+            assert.equal(r.status, 0);
+            console.log(r.output.toString('utf8'));
+            //test included in test-func-env.js
+            assert.equal(!("TEST_HUBID" in process.env), true);
+        });
+    });
+
     describe("* Crashing run", function () {
         it("should fail", function () {
             var command = get_shell("node ../bin/lambda-local -l ./functs/test-func-error.js -e ./events/test-event.js");